Поработал с обработкой ошибок. @ControllerAdvice
