+++ src/main/java/ru/example/BankCard/controller/AccountController.java
@@ -1,5 +1,6 @@
 package ru.example.BankCard.controller;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -8,17 +9,12 @@ import ru.example.BankCard.dto.AccountDto;
 import ru.example.BankCard.service.AccountService;
 
 import java.util.List;
-
+@RequiredArgsConstructor
 @RestController
 @RequestMapping("/accounts")
 public class AccountController {
 
     private final AccountService accountService;
-
-    @Autowired
-    public AccountController(AccountService accountService) {
-        this.accountService = accountService;
-    }
 
     @GetMapping()
     public List<AccountDto> getAccounts() {
