+++ src/main/java/ru/example/BankCard/exception/PersonNotCreateException.java
@@ -1,4 +1,4 @@
-package ru.example.BankCard.util;
+package ru.example.BankCard.exception;
 
 public class PersonNotCreateException extends RuntimeException {
     public PersonNotCreateException(String msg) {
