+++ src/main/java/ru/example/BankCard/repository/AccountsRepository.java
@@ -1,0 +1,7 @@
+package ru.example.BankCard.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.example.BankCard.entity.Account;
+
+public interface  AccountsRepository extends JpaRepository<Account, Integer> {
+}
